requirements.txt
openai
google-generativeai
click
python-dotenv
pyyaml